test: move app_component startup test into tests/
Tests live in tests/, not inline in src/. Moving it out means the
integration crate needs AppComponent.state and SyncService::new_for_test
to be public. Dropping the app cancels its tasks, so the explicit
cancel_all_tasks() call is gone.
